pipeline/requirements.txt
sentence-transformers==2.2.2
chromadb==0.4.22
faiss-cpu==1.7.4
prefect==2.14.10
grpcio==1.60.0
protobuf==4.25.1
python-multipart==0.0.6
sqlalchemy==2.0.25
aiosqlite==0.19.0
boto3==1.34.23
minio==7.2.0
pydantic==2.5.3
pydantic-settings==2.1.0
